# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: API OM - Orientação e Mobilidade
  description: >
    API para gerir professores, alunos, mapas, sessões de treino e análises
    do sistema de Orientação e Mobilidade (OM).

    Todas as rotas protegidas exigem um token JWT no header `Authorization: Bearer <token>`.
  version: 2.0.0
servers:
  - url: http://127.0.0.1:5000/api
    description: Servidor local de desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Mapa:
      type: object
      properties:
        id_mapa:             { type: integer, example: 1 }
        nome_mapa:           { type: string,  example: "Casa Nível 1" }
        caminho_arquivo_xml: { type: string,  example: "/mapas/1_casa_nivel1.xml" }
        caminho_preview:     { type: string,  example: "/previews/1_casa_nivel1.png", nullable: true }
        data_criacao:        { type: string,  example: "2026-01-15 10:30" }
        ativo:               { type: boolean, example: true }
        id_criador:          { type: integer, example: 1 }
        nome_professor:      { type: string,  example: "Prof. Ana Lima" }
    Sessao:
      type: object
      properties:
        id_log:    { type: integer, example: 7 }
        id_mapa:   { type: integer, example: 1 }
        nome_mapa: { type: string,  example: "Casa Nível 1" }
        data:      { type: string,  example: "2026-03-20 14:05" }
...
paths:
...
  /treinos/mapas:
    post:
      tags: [Mapas e Treinos]
      summary: Upload de Novo Mapa
      description: Envia um arquivo de mapa (.xml ou .json) e registra no sistema. Preview (.png/.jpg) é opcional.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [nome_mapa, arquivo_mapa]
              properties:
                nome_mapa:
                  type: string
                  example: "Casa Nível 1"
                arquivo_mapa:
                  type: string
                  format: binary
                  description: Arquivo do mapa (.xml ou .json)
                arquivo_preview:
                  type: string
                  format: binary
                  description: Imagem de preview (.png, .jpg, .jpeg, .webp) — opcional
      responses:
        '201':
          description: Mapa registado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_mapa:         { type: integer }
                  caminho:         { type: string }
                  caminho_preview: { type: string, nullable: true }
        '400': { description: Arquivo ou nome do mapa não informados }
        '403': { description: Apenas professores podem registar mapas }

    get:
      tags: [Mapas e Treinos]
      summary: Listar Todos os Mapas
      description: Retorna todos os mapas do sistema (todos os professores).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de mapas
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  mapas:
                    type: array
                    items: { $ref: '#/components/schemas/Mapa' }
  /treinos/mapas/meus:
    get:
      tags: [Mapas e Treinos]
      summary: Listar Meus Mapas
      description: Retorna apenas os mapas criados pelo professor autenticado.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de mapas do professor
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  mapas:
                    type: array
                    items: { $ref: '#/components/schemas/Mapa' }
  /treinos/mapas/{id_mapa}/ativo:
    patch:
      tags: [Mapas e Treinos]
      summary: Ativar / Desativar Mapa
      description: Alterna ativo/inativo de um mapa do professor autenticado.
      security:
        - bearerAuth: []
      parameters:
        - name: id_mapa
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Novo estado do mapa
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_mapa: { type: integer }
                  ativo:   { type: boolean }
        '404': { description: Mapa não encontrado ou sem permissão }
  /treinos/mapas/{id_mapa}/apropriar:
    post:
      tags: [Mapas e Treinos]
      summary: Apropriar Mapa
      description: Cria uma cópia do registro de um mapa de outro professor para o professor autenticado. Os arquivos são compartilhados.
      security:
        - bearerAuth: []
      parameters:
        - name: id_mapa
          in: path
          required: true
          schema: { type: integer, example: 5 }
      responses:
        '201':
          description: Mapa apropriado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_mapa:   { type: integer }
                  nome_mapa: { type: string }
        '403': { description: Apenas professores podem apropriar mapas }
        '404': { description: Mapa não encontrado }
        '409': { description: Mapa já pertence ao professor ou já foi apropriado }
  /treinos/mapas/{id_mapa}/preview:
    get:
      tags: [Mapas e Treinos]
      summary: Servir Preview do Mapa
      description: Rota pública (sem autenticação) para exibir a imagem de preview de um mapa.
      parameters:
        - name: id_mapa
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Imagem de preview
          content:
            image/png: {}
            image/jpeg: {}
        '404': { description: Preview não disponível }
  /treinos/sessoes:
    post:
      tags: [Mapas e Treinos]
      summary: Registar Sessão de Treino
      description: Faz upload do arquivo de log de uma sessão e o vincula ao aluno e mapa.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [id_aluno, id_mapa, arquivo_log]
              properties:
                id_aluno:    { type: integer, example: 3 }
                id_mapa:     { type: integer, example: 1 }
                arquivo_log:
                  type: string
                  format: binary
                  description: Arquivo de log gerado pelo app (.csv ou .json)
      responses:
        '201':
          description: Sessão registada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_log: { type: integer }
        '403': { description: Aluno não pertence ao professor }
        '404': { description: Aluno ou mapa não encontrado }

    get:
      tags: [Mapas e Treinos]
      summary: Listar Sessões de um Aluno
      security:
        - bearerAuth: []
      parameters:
        - name: id_aluno
          in: query
          required: true
          schema: { type: integer, example: 3 }
      responses:
        '200':
          description: Lista de sessões
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:   { type: integer }
                  sessoes:
                    type: array
                    items: { $ref: '#/components/schemas/Sessao' }
        '400': { description: Parâmetro id_aluno não informado }
        '404': { description: Aluno não encontrado ou sem permissão }
  /treinos/arquivos/{pasta}/{nome_arquivo}:
    get:
      tags: [Mapas e Treinos]
      summary: Baixar Arquivo
      description: Serve arquivos XML, JSON ou de log armazenados no servidor.
      security:
        - bearerAuth: []
      parameters:
        - name: pasta
          in: path
          required: true
          schema:
            type: string
            enum: [mapas, sessoes, analises]
        - name: nome_arquivo
          in: path
          required: true
          schema: { type: string, example: "1_casa_nivel1.xml" }
      responses:
        '200':
          description: Conteúdo do arquivo
          content:
            application/xml: {}
            application/json: {}
        '403': { description: Pasta não permitida }
        '404': { description: Arquivo não encontrado }

  # ============================================================
  # ATIVIDADES
  # ============================================================
...